+++ packages/xrpl/src/models/ledger/NegativeUNL.ts
@@ -1,5 +1,11 @@
 import BaseLedgerEntry from './BaseLedgerEntry'
 
+/**
+ * The unique id for the nUNL object https://xrpl.org/negativeunl.html#negativeunl-id-format
+ */
+export const NEGATIVE_UNL_ID =
+  '2E8A59AA9D3B5B186B0B9E0F62E6C02587CA74A4D778938E957B6357D364B244'
+
 interface DisabledValidator {
   FirstLedgerSequence: number
   PublicKey: string