+++ packages/xrpl/HISTORY.md
@@ -14,7 +14,7 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 * Add type for `account_info`'s `account_flags` property.
 * Add types for `EnableAmendment`, `SetFee`, and `UNLModify` transactions.
 * Add the new fields for `XRPFees` amendment and id for the `FeeSettings`
-* Add `FeeSettings` singleton ledger entry id.
+* Add `FeeSettings`, `NegativeUNL`, and `Amendments` singleton ledger entry ids.
 * Add `WalletLocator` to `SignerEntry` on `SignerList` (LedgerEntry).
 
 ## 2.8.0 (2023-06-13)